improving display of message details

+++ public/app/widget/HermesDetailsWidget.js
@@ -10,10 +10,11 @@ define([
     "dojo/_base/array",
     "dojo/store/Memory",
     "dojox/html/entities",
-    "dojo/topic"
+    "dojo/topic","dojo/dom-style"
 
 
-], function (declare, _WidgetBase, _OnDijitClickMixin, _TemplatedMixin, _WidgetsInTemplateMixin, template, request, Dialog, array, Store, entities, topic) {
+
+], function (declare, _WidgetBase, _OnDijitClickMixin, _TemplatedMixin, _WidgetsInTemplateMixin, template, request, Dialog, array, Store, entities, topic, domStyle) {
     return declare([_WidgetBase, _OnDijitClickMixin, _TemplatedMixin, _WidgetsInTemplateMixin], {
         // Note: string would come from dojo/text! plugin in a 'proper' dijit
         templateString: template,
@@ -66,35 +67,43 @@ define([
         },
 
         _onDeleteClick: function() {
-            var rowsToDelete = this.messagesGridWidget.select.row.getSelected();
 
-            var widget = this;
-            var reloadMessages = this._reloadMessages;
-            request("/services/environnements/" + this.env + "/brokers/" + this.broker + "/queues/" + this.queueName+"/messages/"+rowsToDelete, {"method" : "DELETE"}).then(
-                function (data) {
-                    widget._reloadMessages();
+           var rowsToDelete = this.messagesGridWidget.select.row.getSelected();
 
-                },
-                function (error) {
-                    console.log(error);
+           if(rowsToDelete.length == 0) {
+               alert("There is no message selected");
+           }  else {
+                if(confirm("You are about to delete "+rowsToDelete.length+" msgs from the queue "+this.queueName)) {
+                    var widget = this;
+                    var reloadMessages = this._reloadMessages;
+                    request("/services/environnements/" + this.env + "/brokers/" + this.broker + "/queues/" + this.queueName+"/messages/"+rowsToDelete, {"method" : "DELETE"}).then(
+                        function (data) {
+                            widget._reloadMessages();
 
+                        },
+                        function (error) {
+                            console.log(error);
+
+                        }
+                    );
                 }
-            );
+           }
         },
         _onPurgeClick: function() {
 
-            var widget = this;
 
-            request("/services/environnements/" + this.env + "/brokers/" + this.broker + "/queues/" + this.queueName+"/messages/all", {"method" : "DELETE"}).then(
-                function (data) {
-                    widget._reloadMessages();
+            if(confirm("You are about to purge the queue "+this.queueName)) {
+                var widget = this;
+                request("/services/environnements/" + this.env + "/brokers/" + this.broker + "/queues/" + this.queueName+"/messages/all", {"method" : "DELETE"}).then(
+                    function (data) {
+                        widget._reloadMessages();
+                    },
+                    function (error) {
+                        console.log(error);
 
-                },
-                function (error) {
-                    console.log(error);
-
-                }
-            );
+                    }
+                );
+            }
         },
         _onFileUploadComplete: function( data) {
             this.baseWidgetId._onRefreshClick();
@@ -102,7 +111,6 @@ define([
         _onRowClick : function(evt) {
             var cell = this.messagesGridWidget.cell(evt.rowId, evt.columnId);
             var msg = cell.row.model.byId(cell.row.id).item;
-            console.log(msg);
         //    this.set("detailsContent", entities.encode(msg.content));
             this.set("detailsContent", msg.content);
 
@@ -122,7 +130,6 @@ define([
             array.forEach(rowsToDelete, function(msgId, i){
                 var currentMsg = messagesGridWidget.model.byId(msgId).item;
                 msgs.push(currentMsg);
-                console.log(currentMsg);
             });
             topic.publish("clipboard/copy",msgs);
 
@@ -131,17 +138,89 @@ define([
         _onPasteClick : function() {
             var widget = this;
             topic.publish("clipboard/action", function(msgList) {
-                request.post("/services/environnements/" + widget.env + "/brokers/" + widget.broker + "/queues/" + widget.queueName+"/messages", {data: {"msgs": JSON.stringify(msgList)}}).then(
-                    function (data) {
-                        alert(data);
-                        widget._onRefreshClick();
 
-                    },
-                    function (error) {
-                        console.log(error);
-                    }
-                );
+                if(msgList.length == 0) {
+                    alert("there is no message in the clipboard");
+                }else
+
+                if(confirm("Your are about to paste "+msgList.length +" msgs to the queue "+widget.queueName)) {
+                    request.post("/services/environnements/" + widget.env + "/brokers/" + widget.broker + "/queues/" + widget.queueName+"/messages", {data: {"msgs": JSON.stringify(msgList)}}).then(
+                        function (data) {
+                            alert(data);
+                            widget._onRefreshClick();
+
+                        },
+                        function (error) {
+                            console.log(error);
+                        }
+                    );
+                }
+
+
             });
+        },
+        _onCopyCurrentClick : function() {
+            if(this.detailsId != "- no row selected -") {
+                var msg = this.messagesGridWidget.model.byId(this.detailsId).item;
+                var msgs = [];
+                msgs.push(msg);
+                topic.publish("clipboard/copy",msgs);
+            }
+        },
+        _onMoveSelection : function() {
+
+            alert("Move selection");
+        },
+
+        detailsState : "default",
+
+        _onPropertiesExpand : function() {
+            if(this.detailsState =="default") {
+                domStyle.set(this.fieldsetProperties, "width", "90%");
+                domStyle.set(this.fieldsetContent, "display", "none");
+                this.set("detailsState", "propertiesMax");
+            }
+        },
+        _onPropertiesContract : function() {
+            if(this.detailsState =="propertiesMax") {
+                domStyle.set(this.fieldsetProperties, "width", "30%");
+                domStyle.set(this.fieldsetContent, "display", "block");
+                this.set("detailsState", "default");
+            }  else if(this.detailsState =="default") {
+                domStyle.set(this.fieldsetProperties, "display", "none");
+                domStyle.set(this.fieldsetContent, "width", "90%");
+                this.set("detailsState", "propertiesMin");
+            }  else if(this.detailsState =="propertiesMin") {
+                domStyle.set(this.fieldsetContent, "width", "60%");
+                domStyle.set(this.fieldsetProperties, "width", "30%");
+                domStyle.set(this.fieldsetContent, "display", "block");
+                this.set("detailsState", "default");
+            }
+        },
+
+        _onContentExpand : function() {
+            if(this.detailsState =="default") {
+                domStyle.set(this.fieldsetContent, "width", "90%");
+                domStyle.set(this.fieldsetProperties, "display", "none");
+                this.set("detailsState", "propertiesMax");
+            }
+        },
+
+        _onContentContract : function(evt) {
+            if(this.detailsState =="propertiesMax") {
+                domStyle.set(this.fieldsetContent, "width", "60%");
+                domStyle.set(this.fieldsetProperties, "display", "block");
+                this.set("detailsState", "default");
+            }  else if(this.detailsState =="default") {
+                domStyle.set(this.fieldsetContent, "display", "none");
+                domStyle.set(this.fieldsetProperties, "width", "90%");
+                this.set("detailsState", "propertiesMin");
+            }  else if(this.detailsState =="propertiesMin") {
+                domStyle.set(this.fieldsetContent, "width", "60%");
+                domStyle.set(this.fieldsetProperties, "width", "30%");
+                domStyle.set(this.fieldsetProperties, "display", "block");
+                this.set("detailsState", "default");
+            }
         }
     });
